third_party/WebKit/LayoutTests/editing/selection/caret-at-bidi-boundary-expected.txt

This test ensures WebKit lets user place caret around bidirectional text properly.

Current offset:
abcאבג
0 1 2 3 5 4 6
נת12
2 3 4 1 4 (There is a bug. Clicking on the left of 12 should result in offset 0.)
לשנת
0 3 2 1 4
aקל12יםd
0 1 6 3 4 5 2 7 8
12קל43ab
4 5 6 3 0 1 2 7 8
abcלשנ
6 1 2 6 5 4 3
Test "abcאבג":
PASS caret is at 0
PASS caret is at 1
PASS caret is at 2
PASS caret is at 3
PASS caret is at 5
PASS caret is at 4
PASS caret is at 6


Test "נת12":
PASS caret is at 2
PASS caret is at 3
PASS caret is at 4
PASS caret is at 1
PASS caret is at 4


Test "לשנת":
PASS caret is at 0
PASS caret is at 3
PASS caret is at 2
PASS caret is at 1
PASS caret is at 4


Test "aקל12יםd":
PASS caret is at 0
PASS caret is at 1
PASS caret is at 6
PASS caret is at 3
PASS caret is at 4
PASS caret is at 5
PASS caret is at 2
PASS caret is at 7
PASS caret is at 8


Test "12קל43ab":
PASS caret is at 4
PASS caret is at 5
PASS caret is at 6
PASS caret is at 3
PASS caret is at 0
PASS caret is at 1
PASS caret is at 2
FAIL caret was at 6 but expected to be at 7
PASS caret is at 7
PASS caret is at 8


Test "abcלשנ":
PASS caret is at 6
PASS caret is at 1
PASS caret is at 2
PASS caret is at 6
PASS caret is at 5
PASS caret is at 4
PASS caret is at 3


PASS successfullyParsed is true

TEST COMPLETE
